Updated files and added 2 additional files to improve functionality and test passing threshold
