+++ lesson_16/objects/objects_app/src/main/java/com/codedifferently/lesson16/Dice/DiceMaterial.java
@@ -1,0 +1,8 @@
+package com.codedifferently.lesson16.Dice;
+
+public enum DiceMaterial {
+  PLASTIC,
+  METAL,
+  WOOD,
+  STONE
+}
