+++ lesson_16/objects/objects_app/src/main/java/com/codedifferently/lesson16/Dice/InvalidRollException.java
@@ -1,0 +1,7 @@
+package com.codedifferently.lesson16.Dice;
+
+public class InvalidRollException extends Exception {
+  public InvalidRollException(String message) {
+    super(message);
+  }
+}
